Created functionality of a confirmation dialog to confirm before deleting data in Registration dashboard

+++ src/app/dashboard/dashboard.component.ts
@@ -8,16 +8,18 @@ import { MatTableDataSource } from '@angular/material/table';
 // Import Services
 import { SignUpService } from '../sign-up.service';
 import { UtilityService } from '../utility.service';
+import { MatDialog } from '@angular/material/dialog';
 
 // Import Models
 import { signUpData } from '../model/signupdata';
+import { ConfirmationdialogComponent } from '../confirmationdialog/confirmationdialog.component';
 
 
 @Component({
   selector: 'app-dashboard',
   templateUrl: './dashboard.component.html',
   styleUrl: './dashboard.component.css',
-  providers: [SignUpService, UtilityService]
+  providers: [SignUpService, UtilityService, MatDialog]
 })
 // export class DashboardComponent implements OnInit, AfterViewInit {
   export class DashboardComponent implements OnInit {
@@ -35,7 +37,8 @@ import { signUpData } from '../model/signupdata';
 
   constructor(  // inject services
     private signUpService: SignUpService,
-    private utilityService: UtilityService
+    private utilityService: UtilityService,
+    private matDialog: MatDialog
   ) {}
 
   // ngAfterViewInit(): void { // lifecycle hook that Angular calls after it has fully initialized the component's view. This is the appropriate place to perform any additional initialization that requires access to the view, such as setting up the paginator and sorter for a table.
@@ -75,16 +78,36 @@ import { signUpData } from '../model/signupdata';
 
   // Function invoked when a row is to be deleted
   deleteRow(id: string) {
-    this.signUpService.deleteSignUp(id).subscribe({
-      next: (res) => {
-        this.dataSource.data = this.dataSource.data.filter((user) => user.id !== id); // filter out the deleted row
-        this.utilityService.showSuccessMessage("Record Deleted Succesfully!");
+
+    // Invoking dialog to confirm if row is to be deleted 
+    const dialogRef = this.matDialog.open(ConfirmationdialogComponent, {  // opens the dialog by passing the component having the dialog
+      data: { message: "Are you sure you want to delete this record?" }
+    });
+
+    // open method created an instance of the MatDialogRef which provides a handle on the open dialog
+    dialogRef.afterClosed().subscribe({
+      next: (res1) => {  // value bound in confirmationdialog.component.html     
+        if (res1) { // if true
+          this.signUpService.deleteSignUp(id).subscribe({ // call delete method in signUpService and return the deleted object
+            next: (res2) => {
+              this.dataSource.data = this.dataSource.data.filter((res3) => res3.id !== res2.id); // filter out the deleted object
+              this.utilityService.showSuccessMessage("Record Deleted Succesfully!");
+            },
+            error: (error) => {
+              console.log(error);
+              this.utilityService.showWarningMessage("Record Could Not Be Deleted!");
+            }
+          });
+        }
+        else {
+          this.utilityService.showWarningMessage("Deletion Cancelled!");
+        }      
       },
       error: (error) => {
         console.log(error);
         this.utilityService.showWarningMessage("Record Could Not Be Deleted!");
       }
-    });
+    });    
   }
   
     // Function invoked when a row is to be edited
